fix(low): 12 low-severity code quality improvements
- Improve all expect() messages with descriptive context (9 instances in cloud.rs)
- Fix spawn_scan_idx bounds: use saturating_sub(1) to prevent out-of-bounds start
- Add #[must_use] annotations to pure functions (all_color_schemes, cycle_color_scheme, default_to_ascii, detect_color_mode_auto)
- Replace hardcoded cell size in estimate_memory_budget with std::mem::size_of
- Unify perf_work_max_s from f32 to f64 for consistency with perf_work_sum_s
- Use Cell::blank_with_bg(None) instead of verbose inline Cell construction in LastFrame
- Reuse row_dirty Vec allocation with resize_with instead of replacing with new vec
- Add catch-all arm in color_to_rgb for future-proofing against new crossterm Color variants
- Reject control characters in parse_user_hex_chars to prevent broken terminal rendering
- Add doc comments explaining u16::MAX sentinel values on Droplet fields
- Add doc comments to DrawCtx, ColumnStatus, MsgChr, CloudConfig structs
- Add doc comment to validate_err explaining terminal-restore side effect

+++ src/interactive.rs
@@ -146,7 +146,7 @@ pub fn run_interactive(cfg: &CloudConfig) -> std::io::Result<()> {
     let mut perf_frames: u64 = 0;
     let mut perf_drawn_frames: u64 = 0;
     let mut perf_work_sum_s: f64 = 0.0;
-    let mut perf_work_max_s: f32 = 0.0;
+    let mut perf_work_max_s: f64 = 0.0;
     let mut perf_pressure_sum: f64 = 0.0;
     let mut perf_pressure_max: f32 = 0.0;
     let mut perf_overshoot_frames: u64 = 0;
@@ -289,7 +289,7 @@ pub fn run_interactive(cfg: &CloudConfig) -> std::io::Result<()> {
                 perf_drawn_frames = perf_drawn_frames.saturating_add(1);
             }
             perf_work_sum_s += work_s as f64;
-            perf_work_max_s = perf_work_max_s.max(work_s);
+            perf_work_max_s = perf_work_max_s.max(work_s as f64);
             perf_pressure_sum += perf_pressure as f64;
             perf_pressure_max = perf_pressure_max.max(perf_pressure);
             if overshoot > 0.0 {
@@ -330,7 +330,7 @@ pub fn run_interactive(cfg: &CloudConfig) -> std::io::Result<()> {
             drawn_ratio * 100.0
         );
         println!("  avg_work_ms: {:.3}", avg_work_ms);
-        println!("  max_work_ms: {:.3}", perf_work_max_s as f64 * 1000.0);
+        println!("  max_work_ms: {:.3}", perf_work_max_s * 1000.0);
         println!(
             "  overshoot_frames: {} ({:.1}%)",
             perf_overshoot_frames,
